--- components/BottomNav.module.css
@@ -1,41 +1,0 @@
-.bottomNav {
-  position: fixed;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  height: 64px;
-  background: hsl(var(--background));
-  border-top: 1px solid hsl(var(--border));
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  padding: 0 16px;
-  z-index: 50;
-}
-
-.navItem {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  gap: 4px;
-  padding: 8px;
-  border: none;
-  background: none;
-  color: hsl(var(--muted-foreground));
-  cursor: pointer;
-  transition: color 0.2s;
-  min-width: 64px;
-}
-
-.navItem span {
-  font-size: 12px;
-}
-
-.navItem:hover {
-  color: hsl(var(--foreground));
-}
-
-.navItem.active {
-  color: hsl(var(--primary));
-}
