+++ styles/globals.css
@@ -1,18 +1,4 @@
-:root {
-  --max-width: 1100px;
-  --border-radius: 12px;
-  --font-mono: ui-monospace, Menlo, Monaco, 'Cascadia Mono', 'Segoe UI Mono',
-    'Roboto Mono', 'Oxygen Mono', 'Ubuntu Monospace', 'Source Code Pro',
-    'Fira Mono', 'Droid Sans Mono', 'Courier New', monospace;
-  --gradient-color-1: #1e3c72;
-  --gradient-color-2: #2a5298;
-  --gradient-color-3: #1c3f60;
-  --gradient-color-4: #2a5298;
-  --star-color: rgba(255, 255, 255, 0.8);
-  --twinkle-color: rgba(255, 255, 255, 1);
-}
-
-@media (prefers-color-scheme: dark) {
+@layer base {
   :root {
     --background: 0 0% 100%;
     --foreground: 224 71.4% 4.1%;
@@ -74,9 +60,6 @@ body {
   max-width: 100vw;
   overflow-x: hidden;
   height: 100%;
-  background: radial-gradient(ellipse at bottom, #1b2735 0%, #090a0f 100%);
-  overflow: hidden;
-  min-height: 100vh;
 }
 
 body {
